# backend/api/swagger.yaml
definitions:
...
  api.DeleteUser.deleteUserRequest:
    properties:
      email:
        type: string
    type: object
...
  api.GetCurrentUser.getCurrentUserResponse:
    properties:
      display_name:
        type: string
      email:
        type: string
      id:
        type: integer
      is_admin:
        type: boolean
    type: object
...
  api.GetUsers.getUsersResponse:
    properties:
      numberOfPages:
        type: integer
      numberOfUsers:
        type: integer
      users:
        items:
          $ref: '#/definitions/model.User'
        type: array
    type: object
...
  api.PostUsers.postUsersRequest:
    properties:
      display_name:
        type: string
      email:
        type: string
      is_admin:
        type: boolean
      password:
        type: string
    type: object
...
  model.Friend:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  model.FriendRequest:
    properties:
      id:
        type: integer
      invitation_date:
        type: string
      name:
        type: string
    type: object
  model.User:
    properties:
      display_name:
        type: string
      email:
        type: string
      id:
        type: integer
      is_admin:
        type: boolean
      password_hash:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /user:
    get:
      consumes:
      - application/json
      description: Retrieve the current user.
      parameters:
      - description: Contains the authorization token.
        in: header
        name: authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetCurrentUser.getCurrentUserResponse'
        "400":
          description: Body contains invalid data.
          schema:
            type: string
        "401":
          description: Unauthorized.
          schema:
            type: string
        "409":
          description: User already exists.
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Retrieve the current user.
      tags:
      - User Management
  /user/friends:
    get:
      consumes:
      - application/json
      description: Retrieve the friends of a user.
      parameters:
      - description: Contains the authorization token.
        in: header
        name: authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Friend'
            type: array
        "401":
          description: Unauthorized.
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Retrieve the friends of a user.
      tags:
      - User Management
  /user/friends/requests:
    get:
      consumes:
      - application/json
      description: Retrieve the friend requests of a user.
      parameters:
      - description: Contains the authorization token.
        in: header
        name: authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.FriendRequest'
            type: array
        "401":
          description: Unauthorized.
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Retrieve the friend requests of a user.
      tags:
      - User Management
    post:
      consumes:
      - application/json
      description: send/accept a friend request from the current user to another.
        When the target user does not exist, nothing happens.
      parameters:
      - description: Contains the email of the target user.
        in: body
        name: targetUser
        required: true
        schema:
          type: integer
      - description: Contains the authorization token.
        in: header
        name: authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Cannot post a request to yourself.
          schema:
            type: string
        "401":
          description: Unauthorized.
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: send/accept a friend request from the current user to another.
      tags:
      - User Management
  /user/friends/requests/{targetUser}:
    post:
      consumes:
      - application/json
      description: send/accept a friend request from the current user to another.
        When the target user does not exist, nothing happens.
      parameters:
      - description: Contains the id of the target user.
        in: path
        name: targetUser
        required: true
        type: integer
      - description: Contains the authorization token.
        in: header
        name: authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Cannot post a request to yourself.
          schema:
            type: string
        "401":
          description: Unauthorized.
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: send/accept a friend request from the current user to another.
      tags:
      - User Management
  /users:
    delete:
      consumes:
      - application/json
      description: Deletes a user.
      parameters:
      - description: Contains the identifier for the to be deleted user.
        in: body
        name: user_data
        required: true
        schema:
          $ref: '#/definitions/api.DeleteUser.deleteUserRequest'
      - description: Contains the authorization token.
        in: header
        name: authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Body contains invalid data.
          schema:
            type: string
        "401":
          description: Unauthorized.
          schema:
            type: string
        "404":
          description: User does not exist.
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Deletes a user.
      tags:
      - User Management
    get:
      consumes:
      - application/json
      description: Retrieve all users with pagination.
      parameters:
      - description: Current page for the pagination.
        in: query
        name: page
        required: true
        type: integer
      - description: The size for the pages.
        in: query
        name: page_size
        required: true
        type: integer
      - description: Contains the authorization token.
        in: header
        name: authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetUsers.getUsersResponse'
        "400":
          description: The page parameter cannot exceed the available amount of pages.
          schema:
            type: string
        "401":
          description: Unauthorized.
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Retrieve all users with pagination.
      tags:
      - User Management
    post:
      consumes:
      - application/json
      description: Creates a new user.
      parameters:
      - description: Contains the information for the new user.
        in: body
        name: user_data
        required: true
        schema:
          $ref: '#/definitions/api.PostUsers.postUsersRequest'
      - description: Contains the authorization token.
        in: header
        name: authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Body contains invalid data.
          schema:
            type: string
        "401":
          description: Unauthorized.
          schema:
            type: string
        "409":
          description: User already exists.
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Creates a new user.
      tags:
      - User Management
swagger: "2.0"